Extract class Event
