+++ src/Queue.cpp
@@ -3,7 +3,6 @@
 #include "Queue.h"
 #include "Args.h"
 #include "TimeInfo.h"
-#include <cassert>
 
 void Queue::synced() {
   // log("synced %u %u\n", u32(events.size()), u32(pendingWrite.size()));
@@ -68,28 +67,4 @@ void Queue::finish() {
 
   ::finish(get());
   synced();
-}
-
-Event::Event(EventHolder&& e, TimeInfo* tInfo) :
-  event{std::move(e)},
-  tInfo{tInfo}
-{
-  assert(tInfo);
-}
-
-Event::~Event() {
-  if (event) {
-    [[maybe_unused]] bool finalized = isDone();
-    assert(finalized);
-  }
-}
-
-bool Event::isDone() const {
-  if (!isFinalized) {
-    if (getEventInfo(event.get()) == CL_COMPLETE) {
-      tInfo->add(getEventNanos(get()));
-      isFinalized = true;
-    }
-  }
-  return isFinalized;
 }
